% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.copula.sim.R
\name{compare.copula.sim}
\alias{compare.copula.sim}
\title{Performing the comparison between empirical data and multiple simulated datasets.}
\usage{
compare.copula.sim(object)
}
\arguments{
\item{object}{A copula.sim object for the comparison.}
}
\value{
Returned the comparison of marginal parameter and covariance.
\enumerate{
\item mean.comparison: comparison between empirical marginal mean and average value of simulated marginal mean.
(1) simu.mean: average value of simulated mean
(2) simu.sd: average value of simulated standard error
(3) simu.mean.low.lim: lower limit of 95\% percentile confidence interval
(4) simu.mean.upp.lim: upper limit of 95\% percentile confidence interval
(5) simu.mean.RB: relative bias
(6) simu.mean.SB: standardized bias
(7) simu.mean.RMSE: root mean square error
\item cov.comparison: comparison between empirical covariance and average value of simulated covariance
}
}
\description{
Performing the comparison between empirical data and multiple simulated datasets.
}
\author{
Pei-Shan Yen, Xuemin Gu
}
