% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.sim.R
\name{copula.sim}
\alias{copula.sim}
\title{The function to simulate Copula datasets}
\usage{
copula.sim(
  data.input,
  id.vec,
  arm.vec,
  n.patient,
  n.simulation,
  seed = NULL,
  validation.type = "none",
  validation.sig.lvl = 0.05,
  std.norm.lb = -3,
  std.norm.ub = 3,
  rmvnorm.matrix.decomp.method = "svd",
  verbose = TRUE
)
}
\arguments{
\item{data.input}{The empirical data that you aimed to simulate.}

\item{id.vec}{The identifications for each observation of the input data.}

\item{arm.vec}{The column for identify the arm of the clinical trial.}

\item{n.patient}{The targeted number of patients in simulated data.}

\item{n.simulation}{The number of simulated datasets.}

\item{seed}{The random seed. Default is NULL to use the current seed.}

\item{validation.type}{A string to specify the hypothesis test used to detect the difference
between input data and the simulated data. Default is "none". Other possible methods are
energy distance ("energy"), ball divergence ("ball") and the equability of dependence
("equalCopula", TwoCop package is needed, so you need to install it from CRAN archive.).}

\item{validation.sig.lvl}{The significant level (alpha) value for the hypothesis test.}

\item{std.norm.lb}{The reasonable minimum of standardized normal value.}

\item{std.norm.ub}{The reasonable maximum of standardized normal value.}

\item{rmvnorm.matrix.decomp.method}{The method to do the matrix decomposition used in the function \code{rmvnorm}.}

\item{verbose}{A boolean value to specify whether to print message.}
}
\value{
A copula.sim object with three elements.
\enumerate{
\item data.input: original data; empirical data; historical data
\item data.transform: quantile transformation of data.input
\item data.simul: simulated data
}
}
\description{
The function to simulate Copula datasets
}
