% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.sim.R
\name{data.diff.test}
\alias{data.diff.test}
\title{The function to perform the hypothesis on the input data and the simulated data}
\usage{
data.diff.test(x, y, test.method)
}
\arguments{
\item{x}{A numeric matrix.}

\item{y}{A numeric matrix which is compared to \code{x}.}

\item{test.method}{A string to specify the hypothesis test used to detect the difference
between input data and the simulated data. Other possible methods are
energy distance ("energy"), ball divergence ("ball") and the equability of dependence
("equalCopula", TwoCop package is needed, so you need to install it from CRAN archive.).}
}
\value{
A list with two elements.
\enumerate{
\item p.value: the p-value of the hypothesis test.
\item test.result: the returned object of the hypothesis test.
}
}
\description{
The function to perform the hypothesis on the input data and the simulated data
}
