% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf.inv.R
\name{ecdf.inv}
\alias{ecdf.inv}
\title{Get the inverse of ECDF}
\usage{
ecdf.inv(x, p, sort.flag = TRUE)
}
\arguments{
\item{x}{A vector of numbers which is the original dataset.}

\item{p}{A vector of numbers which is the probability of the new data.}

\item{sort.flag}{A boolean value to specify whether to sort the output data.}
}
\value{
The inverse values of \code{p} based on ECDF of \code{x}.
}
\description{
Get the inverse of ECDF
}
\examples{
ecdf.inv(0:10, c(0.25, 0.75))
ecdf.inv(0:10, c(0.25, 0.75), FALSE)
}
