% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.sim.R
\name{extract.data.sim}
\alias{extract.data.sim}
\title{Converting data.simul in a copula.sim object into a list of wide-form matrices}
\usage{
extract.data.sim(object)
}
\arguments{
\item{object}{A copula object.}
}
\value{
A list of matrices for simulated data.
}
\description{
Converting data.simul in a copula.sim object into a list of wide-form matrices
}
