% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.copula.sim.R
\name{compare.copula.sim}
\alias{compare.copula.sim}
\title{Performing the comparison between empirical data and multiple simulated datasets.}
\usage{
compare.copula.sim(object)
}
\arguments{
\item{object}{A copula.sim object for the comparison.}
}
\value{
Returned the comparison of marginal mean and covariance
}
\description{
Performing the comparison between empirical data and multiple simulated datasets.
}
\author{
Pei-Shan Yen, Xuemin Gu
}
