% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.copula.sim.R
\name{compare.copula.sim}
\alias{compare.copula.sim}
\title{The function to do the comparison between input dataset and simulated datasets.}
\usage{
compare.copula.sim(object)
}
\arguments{
\item{object}{A copula.sim object for the comparison.}
}
\value{
Returned simulated data
}
\description{
The function to do the comparison between input dataset and simulated datasets.
}
