% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.sim.R
\name{extract.data.sim}
\alias{extract.data.sim}
\title{The function to convert data.simul in a copula.sim object into a list of matrices}
\usage{
extract.data.sim(object)
}
\arguments{
\item{object}{A copula object.}
}
\value{
A list of matrices which is simulated data.
}
\description{
The function to convert data.simul in a copula.sim object into a list of matrices
}
