% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.arm.copula.sim.R
\name{new.arm.copula.sim}
\alias{new.arm.copula.sim}
\title{The function to simulate Copula datasets with new arms}
\usage{
new.arm.copula.sim(
  data.input,
  id.vec,
  arm.vec,
  shift.vec.list,
  n.patient,
  n.simulation,
  seed = NULL,
  validation.type = "none",
  validation.sig.lvl = 0.05,
  std.norm.lb = -3,
  std.norm.ub = 3,
  rmvnorm.matrix.decomp.method = "svd",
  verbose = TRUE
)
}
\arguments{
\item{data.input, id.vec, arm.vec, n.patient, n.simulation, seed, std.norm.lb, std.norm.ub, validation.type, validation.sig.lvl, rmvnorm.matrix.decomp.method, verbosee}{Please refer to the function \link{copula.sim}.}

\item{shift.vec.list}{A list of numeric vectors to specify the mean-shifted values for new arms.}
}
\value{
Please refer to the function \link{copula.sim}.
}
\description{
The function to simulate Copula datasets with new arms
}
